requirements.txt
wheel
Flask
mysql-connector-python
requests
uwsgi
smbus
OrangePi.GPIO
gunicorn
python-dotenv
lxml
cssselect
